refactor: remove unnecessary Ignitor class
