+++ src/main.ts
@@ -1,5 +1,7 @@
 import "dotenv/config"
 import "module-alias/register"
-import Ignitor from "@core/Bot/Ignitor"
+import { start } from "@/Core/Bot/start"
 
-new Ignitor().boostrap()
+start()
+  .then(() => console.log("Bot on"))
+  .catch(console.error)
